Wrap all the local changes
I bundled every uncommitted feature from February to April of 2020,
when the project de facto halted.

Nowadays it not of my spirit to lump random blocks of changes without
context or meaning; however, at the end of the development of Racon,
the versioning became really messy and I left a lot of changes
uncommitted because of my then-unproficiency in Git and version control
in general.

Summarizing, the main changes included in this commit are the inclusion
of a debugging system, more test suites, memory and string utilities,
and special types.